+++ lib/core/src/examples.rs
@@ -1,0 +1,2 @@
+pub mod sprite_system;
+pub mod text;
